requirements.txt
click
aiohttp
rich
pydantic
PyJWT
pyyaml
requests
